Refactor: 페이지 컴포넌트를 server component로 유지
