+++ src/app/mj/ReviewModalTest.module.scss
@@ -1,0 +1,18 @@
+.container {
+  width: 100vw;
+  height: 2000rem;
+  padding: 1rem;
+  background-color: lightgray;
+}
+
+.buttons {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  font-size: 2rem;
+}
+
+.test {
+  margin: 2rem;
+  border-color: red;
+}
