+++ src/app/mj/page.module.scss
@@ -1,17 +1,4 @@
 .container {
   width: 100vw;
-  height: 2000rem;
   background-color: lightgray;
-}
-
-.buttons {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  font-size: 2rem;
-}
-
-.test {
-  margin: 2rem;
-  border-color: red;
 }
